+++ CHANGES.md
@@ -8,6 +8,7 @@ History
 * Add XPath support with `Page.xpath`, `Frame.xpath`, and `ElementHandle.xpath`
 * Change `Frame.executionContest` from property to coroutine
 * Add `ignoreDefaultArgs` option to `pyppeteer.launch`
+* Add `Page.setDefaultNavigationTimeout` method
 * `Page.waitFor*` methods accept `JSHandle` as argument
 * Implement `Frame.content` and `Frame.setContent` methods
 * `page.tracing.start` accepts custom tracing categories option